requirements.txt
pandas
numpy 
matplotlib
seaborn
plotly.express
altair>=5
streamlit>=1.30.0